Update config and add template options
This change adds a number of options needed for cluster-api integration.

This is not a propper integration of magnum-cluster-api, as that would
mean a re-architecture of the charm, but we do add some needed options
to tweak options required by magnum-cluster-api such as default-network-driver
and allowed-network-drivers. Other options deal with RBAC defaults and
adding additional CA certs (useful for bundling root CA certs used by
public endpoints).

Signed-off-by: Gabriel Adrian Samfira <gsamfira@cloudbasesolutions.com>